requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.7
DateTime==5.4
discord.py==2.3.2
fonttools==4.47.0
frozenlist==1.4.1
fuzzywuzzy==0.18.0
idna==3.6
iniconfig==2.0.0
kiwisolver==1.4.5
Levenshtein==0.23.0
matplotlib==3.8.2
multidict==6.0.4
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.2.0
pluggy==1.3.0
pyparsing==3.1.1
pytest==7.4.4
python-dateutil==2.8.2
python-Levenshtein==0.23.0
pytz==2023.3.post1
rapidfuzz==3.6.1
setuptools==69.0.3
six==1.16.0
tzdata==2023.4
yarl==1.9.4
zope.interface==6.1
